requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1740481426959/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
array_record==0.7.2
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
av==14.2.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560556141/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1731428302318/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1723662581198/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
distro==1.9.0
dm-tree==0.1.9
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
flash_attn==2.7.4.post1
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1738940303262/work
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1737645236190/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1710808267764/work
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1733462551614/work
googleapis-common-protos==1.70.0
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.7
httpx==0.28.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1741710986684/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
immutabledict==4.2.1
importlib_resources==6.5.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.9.0
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1733736026804/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
llava @ git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@09e5840d5589ad2d6a8656c0a60f21ae134b3309
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
matplotlib==3.10.1
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1733913043842/work
multiprocess @ file:///home/conda/feedstock_root/build_artifacts/multiprocess_1729987567233/work
munkres==1.1.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1739425582012/work/dist/numpy-2.2.3-cp313-cp313-linux_x86_64.whl#sha256=554cd4f0a07dab06413547bc6579bac9be0b9c5f5bb952191ad69f390fabdf07
open_clip_torch @ git+https://github.com/mlfoundations/open_clip.git@c4fba838acbf4dc94736489164868c3c172a7433
openai==1.70.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1741963793705/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1726878403859/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1735929688808/work
platformdirs==4.3.7
promise==2.3
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1737635528546/work
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix @ file:///home/conda/feedstock_root/build_artifacts/pyarrow-hotfix_1734380560621/work
pybind11 @ file:///D:/bld/pybind11-split_1730237329882/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1730237328107/work
pycocotools==2.0.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.2
pydantic_core==2.33.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1735698276679/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
regex @ file:///home/conda/feedstock_root/build_artifacts/regex_1730952226786/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1740651473232/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1736496745920/work/dist/scikit_learn-1.6.1-cp313-cp313-linux_x86_64.whl#sha256=d95aa585d4be3c6fa9f19265162749c371f7a348da2578b142d9a17040c8d5c4
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1739790645292/work/dist/scipy-1.15.2-cp313-cp313-linux_x86_64.whl#sha256=9233a1b3513262c15ef8696233b9abd49f691223e9dcc32da39377a3a0d4e7fe
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
sentencepiece==0.2.0
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==75.8.2
simple-parsing==0.1.7
simsimd==6.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap==5.0.2
sniffio==1.3.1
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1727986692822/work
stringzilla==3.12.4
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1736248176451/work
tensorflow-datasets==4.9.8
tensorflow-metadata==1.17.1
termcolor==3.1.0
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tiktoken==0.9.0
timm==1.0.15
tokenizers @ file:///home/conda/feedstock_root/build_artifacts/tokenizers_1741890291002/work/bindings/python
toml==0.10.2
torch @ file:///home/conda/feedstock_root/build_artifacts/libtorch_1741576595299/work
torchaudio==2.6.0+cu126
torchvision @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_torchvision_1739668599/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1739826231190/work
triton @ file:///home/conda/feedstock_root/build_artifacts/triton_1741774229062/work/python
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1737541069190/work
ultralytics==8.3.127
ultralytics-thop==2.0.14
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
wandb==0.19.9
wcwidth==0.2.13
wrapt==1.17.2
xxhash @ file:///home/conda/feedstock_root/build_artifacts/python-xxhash_1740594782459/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1737575777699/work
zipp==3.21.0
zstandard==0.23.0